+++ core/src/Logger.kt
@@ -4,7 +4,14 @@ package klite
 import java.lang.System.Logger.Level.*
 
 fun logger(name: String): System.Logger = System.getLogger(name)
-fun Any.logger(): System.Logger = System.getLogger(javaClass.name)
+fun Any.logger(): System.Logger = logger(nonAnonymousClassName())
+
+internal fun Any.nonAnonymousClassName(): String {
+  var cls = javaClass
+  while (cls.isAnonymousClass) cls = cls.superclass
+  return cls.name
+}
+
 inline fun System.Logger.debug(msg: String) = log(DEBUG, msg)
 inline fun System.Logger.info(msg: String) = log(INFO, msg)
 inline fun System.Logger.warn(msg: String) = log(WARNING, msg)